backend/requirements.txt
# Core Framework
Django==4.2.4
djangorestframework==3.14.0
django-cors-headers==4.2.0
drf-spectacular==0.26.4

# Database (PostgreSQL only)
psycopg2-binary==2.9.10
dj-database-url==3.0.1
django-extensions==3.2.3

# Task Queue & Caching
celery==5.5.3
redis==6.4.0
kombu==5.5.4

# AI/ML Framework
langchain>=0.3.7
langchain-core>=0.3.17
langchain-openai>=0.2.8
langgraph>=0.2.34
langsmith>=0.1.137
langchain-community>=0.3.7
openai==1.108.0
sentence-transformers==2.2.2
numpy==1.26.4
pandas==2.3.2
scikit-learn==1.7.2
transformers==4.56.1
torch==2.8.0

# Web Scraping
requests==2.32.4
beautifulsoup4==4.13.5
praw==7.8.1
fake-useragent==1.3.0

# Data Processing
nltk==3.8.1
PyYAML==6.0.2

# Security
cryptography==41.0.3
PyJWT==2.8.0

# Utilities
python-dotenv==1.0.0
gunicorn==21.0.1
whitenoise==6.5.0
inflection==0.5.1

# Development
pytest==7.4.0
pytest-django==4.5.2
black==23.7.0
isort==5.12.0
flake8==6.0.0
factory-boy==3.3.0